feat: Optimize the structure of statistical kanban and task execution details
