+++ ui/src/pages/DataAnalysis/StatisticalBoard/components/Screening/index.less
@@ -1,0 +1,8 @@
+.ScreeningBox {
+  width: 100%;
+  height: 50px;
+  padding-left: 310px;
+  display: flex;
+  align-items: center;
+  justify-content: start;
+}
