+++ ui/src/pages/DataAnalysis/StatisticalBoard/index.less
@@ -2,7 +2,6 @@
   position: relative;
   width: 100%;
   height: 100%;
-  padding-top: 50px;
   &::before {
     position: absolute;
     top: 0;
