+++ ui/src/pages/DataAnalysis/TaskExecutionDetails/TaskFilter/index.less
@@ -1,6 +1,7 @@
 .taskFilter {
-    display: flex;
-    align-items: center;
-    line-height: 50px;
-    font-size: 16px;
+  display: flex;
+  align-items: center;
+  padding-left: 20px;
+  font-size: 14px;
+  line-height: 50px;
 }
